requirements.txt
sounddevice==0.4.6
scipy==1.11.4
numpy==1.24.3
openai>=1.0.0
python-dotenv==1.0.0
edge-tts>=6.1.0
